News/requirements.txt
django
djangorestframework
celery
psycopg2
beautifulsoup4
requests